+++ wagtail/search/tests/test_backends.py
@@ -913,15 +913,6 @@ class BackendTests(WagtailTestUtils):
         )
         self.assertFalse(stdout.read())
 
-    def test_rebuild_references_index_no_verbosity(self):
-        stdout = StringIO()
-        management.call_command(
-            "rebuild_references_index",
-            verbosity=0,
-            stdout=stdout,
-        )
-        self.assertFalse(stdout.read())
-
 
 @override_settings(
     WAGTAILSEARCH_BACKENDS={"default": {"BACKEND": "wagtail.search.backends.database"}}
