+++ wagtail/tests/test_reference_index.py
@@ -1,4 +1,7 @@
+from io import StringIO
+
 from django.contrib.contenttypes.models import ContentType
+from django.core import management
 from django.test import TestCase
 
 from wagtail.images import get_image_model
@@ -159,3 +162,12 @@ class TestCreateOrUpdateForObject(TestCase):
             ),
             self.expected_references,
         )
+
+    def test_rebuild_references_index_no_verbosity(self):
+        stdout = StringIO()
+        management.call_command(
+            "rebuild_references_index",
+            verbosity=0,
+            stdout=stdout,
+        )
+        self.assertFalse(stdout.read())
